Changed EnemyBehaviour (and subclass) constructor parameters
Now instead of passing the Enemy as a parameter to EnemyBehaviour subclass constructors, we pass just the Enemy's Rigidbody2D. This means we don't have to call GetComponent<Rigidbody2D> in those constructors, saving computation time.
Further, I've moved the assignment of the _enemyBody instance variable to a new base EnemyBehaviour constructor to make the code cleaner.